Break out router, use ember-data for snapshot data with new API.
